szerver/requirements.txt
Django
django-cors-headers
django-rest-framework
djangorestframework
pytz
tqdm
requests
